Add MCUBoot pip dependencies
It has been a common request for CI to be able to integration test
MCUBoot, and in
https://github.com/zephyrproject-rtos/zephyr/pull/13672 this becomes
possible, but it requires that Zephyr's CI has MCUBoot's dependencies
installed.

To resolve this we add MCUBoot's dependencies to the CI Dockerfile,
currently, this consists only of pip packages, so we download
MCUboot's requirements.txt file and have pip install them.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>